Fixed Test Cases Read below
created a referenceList that contains a valid quote q. this is then compared to the size() and the contents of the testQuote list after it is set to the result of referenceList.search. resulting in either the same list, or one being empty as the author/quote didnt match when searching